Revert "Return start/end profiling time in microseconds instead of milliseconds"
This reverts r16049 for breaking build on windows.

TBR=svenpanne@chromium.org,machenbach@chromium.org

Review URL: https://codereview.chromium.org/22189002

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@16050 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ include/v8-profiler.h
@@ -149,16 +149,16 @@ class V8EXPORT CpuProfile {
   const CpuProfileNode* GetSample(int index) const;
 
   /**
-    * Returns time when the profile recording started (in microseconds
+    * Returns time when the profile recording started (in milliseconds
     * since the Epoch).
     */
-  int64_t GetStartTime() const;
+  double GetStartTime() const;
 
   /**
-    * Returns time when the profile recording was stopped (in microseconds
+    * Returns time when the profile recording was stopped (in milliseconds
     * since the Epoch).
     */
-  int64_t GetEndTime() const;
+  double GetEndTime() const;
 
   /**
    * Deletes the profile and removes it from CpuProfiler's list.
